refactor : 도메인 계층의 모든 자바코드를 코틀린 코드로 리팩토링

+++ src/main/java/com/group/libraryapp/service/user/UserService.java
@@ -41,8 +41,10 @@ public class UserService {
 
   @Transactional
   public void deleteUser(String name) {
-    User user = userRepository.findByName(name).orElseThrow(IllegalArgumentException::new);
-    userRepository.delete(user);
+    User user = userRepository.findByName(name);
+    if (user != null) {
+      userRepository.delete(user);
+    }
   }
 
 }
